+++ web/src/beta/lib/core/StoryPanel/Block/builtin/common/Wrapper.tsx
@@ -3,8 +3,8 @@ import { ReactNode, createContext } from "react";
 import { stopClickPropagation } from "@reearth/beta/utils/events";
 import { styled } from "@reearth/services/theme";
 
+import SelectableArea from "../../../../shared/components/SelectableArea";
 import { FieldComponent } from "../../../hooks/useFieldComponent";
-import SelectableArea from "../../../SelectableArea";
 import Template from "../../Template";
 
 import useHooks from "./hooks";
