+++ web/src/beta/lib/core/shared/components/SelectableArea/ActionPanel/index.tsx
@@ -1,6 +1,6 @@
 import { Dispatch, MouseEvent, SetStateAction } from "react";
 
-import ActionPanel, { type ActionPosition } from "../../../../ActionPanel";
+import ActionPanelUI, { type ActionPosition } from "../../ActionPanel";
 
 import useHooks from "./hooks";
 
@@ -46,37 +46,45 @@ type Props = {
   ) => Promise<void>;
 };
 
-const BlockActionPanel: React.FC<Props> = ({
+const ActionPanel: React.FC<Props> = ({
   title,
   icon,
   isSelected,
   editMode,
   dndEnabled,
   panelSettings,
+  setShowPadding,
   onEditModeToggle,
   onSettingsToggle,
+  onRemove,
   ...actionProps
 }) => {
-  const { actionItems } = useHooks({
+  const { dndItemContextRef, settingsTitle, popoverContent, actionItems } = useHooks({
     title,
     icon,
     isSelected,
     editMode,
     panelSettings,
+    setShowPadding,
     onEditModeToggle,
     onSettingsToggle,
+    onRemove,
   });
 
   return (
-    <ActionPanel
+    <ActionPanelUI
+      ref={dndItemContextRef}
       dndEnabled={dndEnabled}
       isSelected={isSelected}
       actionItems={actionItems}
       panelSettings={panelSettings}
+      settingsTitle={settingsTitle}
+      popoverContent={popoverContent}
+      setShowPadding={setShowPadding}
       onSettingsToggle={onSettingsToggle}
       {...actionProps}
     />
   );
 };
 
-export default BlockActionPanel;
+export default ActionPanel;
